Fixes crash occurring when adding photo in Photo Library while selecting photos to upload

+++ piwigo/Albums/AlbumImagesViewController.m
@@ -1449,7 +1449,7 @@ NSString * const kPiwigoNotificationDeletedImage = @"kPiwigoNotificationDeletedI
         
         // Image Id?
         NSInteger imageId = [[userInfo objectForKey:@"imageId"] integerValue];
-        NSLog(@"=> removeImage %ld to Category %ld", (long)imageId, catId);
+        NSLog(@"=> removeImage %ld to Category %ld", (long)imageId, (long)catId);
         
         // Store current image list
         NSArray *oldImageList = self.albumData.images;
